new changes

+++ HamiSabaiko.Entities/Customer.cs
@@ -10,7 +10,7 @@ namespace HamiSabaiko.Entities
     {
         public int SerialNumber { get; set; }
         public string Name { get; set; }
-        public decimal Amount { get; set; }
-        public decimal LoanAmount { get; set; }
+        public int Amount { get; set; }
+        public int LoanAmount { get; set; }
     }
 }
